requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0
scikit-learn>=1.3.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
pyarrow>=12.0.0

# Hugging Face
huggingface-hub>=0.16.0

# Utilities
tqdm>=4.65.0
